% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reference_data.R
\name{download_reference_data}
\alias{download_reference_data}
\title{Download reference files}
\usage{
download_reference_data()
}
\description{
Downloads needed reference files from \url{https://alkesgroup.broadinstitute.org/LDSCORE/} and \url{https://storage.googleapis.com/broad-alkesgroup-public/LDSCORE/eur_w_ld_chr.tar.bz2}
}
\examples{
NA
}
