% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{est_enrichment}
\alias{est_enrichment}
\title{Enrichment in genomic regions of a variance of a common latent factor}
\usage{
est_enrichment(traits, gen_cov_strat)
}
\arguments{
\item{traits}{Traits making up the common latent factor, should beat least 3}

\item{gen_cov_strat}{Stratified genetic covariance matrices}
}
\value{
Tibble of enrichment in regulatory genomic regions
}
\description{
Enrichment in genomic regions of a variance of a common latent factor
}
\examples{
NA
}
