% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_covariance.R
\name{est_gen_cov}
\alias{est_gen_cov}
\title{Stratified genetic covariance between traits using LDScore regression}
\usage{
est_gen_cov(paths, sample_prev, population_prev, trait_names)
}
\arguments{
\item{paths}{Paths to munged summary statistics}

\item{sample_prev}{Sample prevalences for GWASs. NA for continuous}

\item{population_prev}{Population prevalences for traits. NA for continuous}

\item{trait_names}{Specify names for traits}
}
\value{
Genetic covariance object estimated from GenomicSEM::ldsc
}
\description{
Stratified genetic covariance between traits using LDScore regression
}
\examples{
NA
}
